requirements.txt
dash>=2.1.0
dash_bootstrap_components>=1.0.3
diff_match_patch
numpy
plotly
SoundFile
tqdm
prefect
datasets
jiwer
matplotlib
google-cloud-speech
azure-cognitiveservices-speech
openai
openai-whisper
transformers
librosa
Cython
nemo_toolkit
hydra-core
torch
seaborn
pandas
pytorch-lightning
sentencepiece
youtokentome
inflect
webdataset
pyannote.core
pyannote.metrics
editdistance
ipython
assemblyai
argilla
huggingface-hub==0.23.2
pydub
git+https://github.com/espnet/espnet
espnet_model_zoo
torchaudio --index-url https://download.pytorch.org/whl/cu118
torch --index-url https://download.pytorch.org/whl/cu118
wheel
ninja
langdetect